pyproject.toml
[project]
name = "12labs"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "twelvelabs>=1.2.2",
]
